+++ src/documents/management/commands/document_importer.py
@@ -94,7 +94,9 @@ class Command(Renderable, BaseCommand):
 
             document_path = os.path.join(self.source, doc_file)
             thumbnail_path = os.path.join(self.source, thumb_file)
-            thumbnail_webp_path = os.path.join(self.source, thumb_webp_file)
+            
+            if thumb_webp_file:
+                thumbnail_webp_path = os.path.join(self.source, thumb_webp_file)
 
             if settings.PASSPHRASE:
 
@@ -110,17 +112,20 @@ class Command(Renderable, BaseCommand):
                             thumb_file, document.thumbnail_path))
                         encrypted.write(GnuPG.encrypted(unencrypted))
 
-                with open(thumbnail_webp_path, "rb") as unencrypted:
-                    with open(document.thumbnail_webp_path, "wb") as encrypted:
-                        print("Encrypting {} and saving it to {}".format(
-                            thumb_webp_file, document.thumbnail_path_webp))
-                        encrypted.write(GnuPG.encrypted(unencrypted))
+                if os.path.exists(thumbnail_webp_path):
+                    with open(thumbnail_webp_path, "rb") as unencrypted:
+                        with open(document.thumbnail_webp_path, "wb") as encrypted:
+                            print("Encrypting {} and saving it to {}".format(
+                                thumb_webp_file, document.thumbnail_path_webp))
+                            encrypted.write(GnuPG.encrypted(unencrypted))
 
             else:
 
                 shutil.copy(document_path, document.source_path)
                 shutil.copy(thumbnail_path, document.thumbnail_path)
-                shutil.copy(thumbnail_webp_path, document.thumbnail_path_webp)
+
+                if os.path.exists(thumbnail_webp_path):
+                    shutil.copy(thumbnail_webp_path, document.thumbnail_path_webp)
 
         # Reset the storage type to whatever we've used while importing
